+++ agent/modals/ticket/ticket_edit.php
@@ -2,9 +2,9 @@
 
 require_once '../../../includes/modal_header.php';
 
-// Ticket client access overide
+// Ticket client access overide - This is the only way to show tickets without a client to agents with restricted client access
 $access_permission_query_overide = '';
-if (!empty($client_access_string)) {
+if ($client_access_string) {
     $access_permission_query_overide = "AND ticket_client_id IN (0,$client_access_string)";
 }